CAS-180 Move initials extensions to clinet module
