+++ stream-chat-android/src/main/java/com/getstream/sdk/chat/utils/Primitives.kt
@@ -5,10 +5,3 @@ internal inline fun Boolean.whenFalse(crossinline f: () -> Unit): Boolean = also
 
 internal inline val <T> T.exhaustive: T
     get() = this
-
-internal fun String?.initials(): String =
-    this
-        ?.trim()
-        ?.split("\\s+".toRegex())
-        ?.take(2)?.joinToString(separator = "") { it.take(1).toUpperCase() }
-        ?: ""
